update shell script

+++ Tools/scripts/build_align.sh
@@ -2,13 +2,17 @@
 # build for all copters boards
 
 set -e
-set -x
 
-rm -rf align-build
-mkdir align-build
-
+# specify boards
 BOARDS="AP3-M460 AP3-M460-dev AP3-M460-G3P"
 
+# prepare align-build folder
+if [ -d "align-build" ]; then
+    rm -rf align-build
+fi
+mkdir align-build
+
+# compile
 for b in $BOARDS; do
     echo "Testing $b build"
     ./waf configure --board $b
